// src/app/models/product.model.ts
// export interface Product {
//   id: number;
//   title: string;
//   price: number;
//   category: string;
//   description: string;
//   imagePath: string;
// }
// image:string;
// product.model.ts

export interface Product {
  id: number;
  name: string;
  category: string;
  description: string;
  price: number;
  image: string;
}